backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy==2.0.36

# Vector store & embeddings
faiss-cpu==1.9.0.post1
google-generativeai==0.8.3

# Document processing
langchain==0.3.13
langchain-text-splitters==0.3.4
pypdf==5.1.0
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5

# OpenRouter (OpenAI-compatible)
openai==1.58.1

# Configuration
pydantic-settings==2.7.0
python-dotenv==1.0.1

# Utilities
aiofiles==24.1.0
